added code to scale the rate measurement based on sensitivity S0

+++ sctebot_sensors/adafruit_9d0f_imu/src/adafruit_9dof_imu.cpp
@@ -27,8 +27,8 @@ void handle_bmp180_measurements(float temperature, float pressure) {
      /**/
 }
 
-void handle_l3gd20_measurements(int temperature, int r_x, int r_y, int r_z) {
-    std::cout << "l3gd20\t\ttemp: " << (float)temperature << " x_dps: " << (float)r_x << " y_dps: " << (float)r_y << " z_dps: " << (float)r_z << std::endl;
+void handle_l3gd20_measurements(int temperature, float r_x, float r_y, float r_z) {
+    std::cout << "l3gd20\t\ttemp: " << temperature << " x_dps: " << r_x << " y_dps: " << r_y << " z_dps: " << r_z << std::endl;
 }
 
 void handle_lsm303dlhc_measurements(int temperature, int a_x, int a_y, int a_z, int m_x, int m_y, int m_z) {
